server/requirements.txt
django              == 2.0.5
djangorestframework == 3.8.2
django-environ      == 0.4.4
django-extensions   == 2.0.7
django-cors-headers == 2.2.0

requests            >=2.18.4, <3

sawtooth-sdk        == 1.0.4

arrow               == 0.12.1
jsonschema          == 2.5.1